created createCart mutation and adding access role security

+++ db/migrate/20190119173036_create_carts.rb
@@ -1,7 +1,7 @@
 class CreateCarts < ActiveRecord::Migration[5.2]
   def change
     create_table :carts do |t|
-      t.decimal :subtotal
+      t.float :subtotal
       t.references :cart_status, foreign_key: true
 
       t.timestamps
