+++ db/migrate/20190119173127_create_cart_items.rb
@@ -3,7 +3,7 @@ class CreateCartItems < ActiveRecord::Migration[5.2]
     create_table :cart_items do |t|
       t.references :product, foreign_key: true, on_delete: :cascade
       t.references :cart, foreign_key: true, on_delete: :cascade
-      t.decimal :unit_price
+      t.float :unit_price
 
       t.timestamps
     end
